Add imu_tools launch files and imu_mag_repacking.py
The MTi-100 does not publish orientations, so a filter must be used to
estimate them. Added launch files with madgwick filter and complementary
filter support.

The xsens drivers and madgwick filter use different message types for the
magnetometer measurements. The `imu_mag_repacking.py` converts a
`Vector3Stamped` to a `MagneticField` message.